Move testing function to a separate file
Change the testing function so that both fcm_get() and fcm_next()
are being tested regardless of the specified parameters/variant.

In preparation for the split variants (each of the three main
implementation variants in a separate file but with a common
header fcmat.h), change the order of evaluation of the variable
args so that FCM_THREAD is now evaluated before FCM_CACHE.

Move the testing function from the main function in fcmat.c
to the new file test_fcmat.c.