Fix: set the appropriate element retrieval function
In case of the half-stored variant, the appropriate element
retrieval function has been set only for iteration (fcm_next)
but not for regular access (fcm_get). As a result, the half-
stored variant performed poorly if fcm_get() was used because
each element was computed twice, once during fcm_create()
and once upon access (using one of the on-demand retrieval
functions).