Make the application of Fisher's r-to-z transformation more efficient
In the case of the half-stored variant, Fisher's r-to-z transformation was
applied after the correlations had already been computed by passing through
the elements again (no multi-threading). I have now integrated the
transformation into a new retrieval function. The transformation is now
applied upon accessing the elements, which should increase the performance
since the accesses typically occur through multi-threaded functions.
A disadvantage is that, in the case of accessing the same element more
than once, the transformation is now recomputed every time. Perhaps we
should add an option to the corr functions that allows for the computation
of the transformation during the computation of the correlation matrix.