Improve corr. computation in fcm_corr_single() and fcm_corr_wrk()
Avoid using pccx() because:

(1) If pccx() is called for every pair, the add. variable is
normalized every time instead of being normalized only once in
the beginning.

(2) To choose the optimal parameters for pccx() in this case
(don't use tiling for just one pair), it is currently necessary to
find out _here_ which instruction set extension should be used.